improved feat taking and legality checks

+++ Patherpoint/src/app/Character.ts
@@ -7,6 +7,7 @@ import { CharacterService } from './character.service';
 import { SkillChoice } from './SkillChoice';
 import { LoreChoice } from './LoreChoice';
 import { AbilityChoice } from './AbilityChoice';
+import { FeatChoice } from './FeatChoice';
 
 export class Character {
     public name: string = "";
@@ -22,7 +23,7 @@ export class Character {
     set_Changed(characterService: CharacterService, ) {
         characterService.set_Changed();
     }
-    get_AbilityBoosts(minLevelNumber: number, maxLevelNumber: number, abilityName: string = "", source: string = "", type: string = "", locked: boolean = undefined ) {
+    get_AbilityBoosts(minLevelNumber: number, maxLevelNumber: number, abilityName: string = "", type: string = "", source: string = "", sourceId: number = -1, locked: boolean = undefined ) {
         if (this.class) {
             let boosts = [];
             let levels = this.class.levels.filter(level => level.number >= minLevelNumber && level.number <= maxLevelNumber );
@@ -30,8 +31,9 @@ export class Character {
                 level.abilityChoices.forEach(choice => {
                     choice.boosts.filter(boost => 
                         (boost.name == abilityName || abilityName == "") &&
-                        (boost.source == source || source == "") &&
                         (boost.type == type || type == "") &&
+                        (boost.source == source || source == "") &&
+                        (boost.sourceId == sourceId || sourceId == -1) &&
                         (boost.locked == locked || locked == undefined)
                     ).forEach(boost => {
                         boosts.push(boost);
@@ -70,8 +72,14 @@ export class Character {
         newChoice.id += 100 + existingChoices.length;
         let newId: number = level.loreChoices.push(newChoice);
         return level.loreChoices[newId-1];
-    } 
-    get_SkillIncreases(minLevelNumber: number, maxLevelNumber: number, skillName: string = "", source: string = "", locked: boolean = undefined) {
+    }
+    add_FeatChoice(level, newChoice: FeatChoice) {
+        let existingChoices = level.featChoices.filter(choice => choice.source == newChoice.source);
+        newChoice.id += existingChoices.length;
+        let newId: number = level.featChoices.push(newChoice);
+        return level.featChoices[newId-1];
+    }
+    get_SkillIncreases(minLevelNumber: number, maxLevelNumber: number, skillName: string = "", source: string = "", sourceId: number = -1, locked: boolean = undefined) {
         if (this.class) {
             let increases = [];
             let levels = this.class.levels.filter(level => level.number >= minLevelNumber && level.number <= maxLevelNumber );
@@ -80,6 +88,7 @@ export class Character {
                     choice.increases.filter(increase => 
                         (increase.name == skillName || skillName == "") &&
                         (increase.source == source || increase.source.indexOf(source) > -1 || source == "") &&
+                        (increase.sourceId == sourceId || sourceId == -1) &&
                         (increase.locked == locked || locked == undefined)
                         ).forEach(increase => {
                         increases.push(increase);
@@ -89,6 +98,7 @@ export class Character {
                     choice.increases.filter(increase => 
                         (increase.name == skillName || skillName == "") &&
                         (increase.source == source || increase.source.indexOf(source) > -1 || source == "") &&
+                        (increase.sourceId == sourceId || sourceId == -1) &&
                         (increase.locked == locked || locked == undefined)
                         ).forEach(increase => {
                         increases.push(increase);
@@ -112,112 +122,40 @@ export class Character {
         }
         this.set_Changed(characterService);
     }
-    get_FeatsTaken(minLevelNumber: number, maxLevelNumber: number, featName: string = "", source: string = "") {
+    get_FeatsTaken(minLevelNumber: number, maxLevelNumber: number, featName: string = "", source: string = "", sourceId: number = -1, locked: boolean = undefined) {
         if (this.class) {
             let featsTaken = [];
             let levels = this.class.levels.filter(level => level.number >= minLevelNumber && level.number <= maxLevelNumber );
             levels.forEach(level => {
-                level.feats.filter(feat => (feat.name == featName || featName == "") && (feat.source == source || source == "")).forEach(feat => {
-                    featsTaken.push(feat);
+                level.featChoices.forEach(choice => {
+                    choice.feats.filter(feat => 
+                        (feat.name == featName || featName == "") &&
+                        (feat.source == source || source == "") &&
+                        (feat.sourceId == sourceId || sourceId == -1) &&
+                        (feat.locked == locked || locked == undefined)
+                        ).forEach(feat => {
+                        featsTaken.push(feat);
+                    })
                 })
             })
             return featsTaken;
         }
     }
-    takeFeat(characterService: CharacterService, level: Level, featName: string, type: string, take: boolean, source: string) {
-        if (take) {
-            level.feats.push({"name":featName, "source":source});
-            let feat = characterService.get_Feats(featName);
-            if (feat.length > 0 && feat[0].increase) {
-                let newSkillChoice = this.add_SkillChoice(level, {available:0, increases:[], type:"Any", maxRank:2, source:'Feat: '+featName, id:0});
-                this.increase_Skill(characterService, feat[0].increase, true, newSkillChoice, true);
-                //If a feat trains you in a skill you don't already have, it's usually a weapon proficiency
-                //We have to create that skill here then
-                if (characterService.get_Skills(feat[0].increase).length == 0 ) {
-                    characterService.add_CustomSkill(feat[0].increase, "Specific Weapon Proficiency", "")
-                }
-            }
-            if (source == "Level") {
-                switch (type) {
-                    case "General":
-                        level.generalFeats_applied += 1;
-                        break;
-                    case "Class":
-                        level.classFeats_applied += 1;
-                        break;
-                    case "Skill":
-                        level.skillFeats_applied += 1;
-                        break;
-                    case "Ancestry":
-                        level.ancestryFeats_applied += 1;
-                        break;
-                }
-            }
-            if (feat.length > 0 && feat[0].gainAncestryFeat) {
-                characterService.get_Character().class.levels.filter(level => level.number == 1)[0].ancestryFeats_available += 1;
-            }
-            if (feat.length > 0 && feat[0].gainGeneralFeat) {
-                characterService.get_Character().class.levels.filter(level => level.number == 1)[0].generalFeats_available += 1;
-            }
-            if (feat.length > 0 && feat[0].gainClassFeat) {
-                characterService.get_Character().class.levels.filter(level => level.number == 1)[0].classFeats_available += 1;
-            }
-            if (feat.length > 0 && feat[0].gainSkillTraining) {
-                this.add_SkillChoice(level, {available:1, increases:[], type:"Skill", maxRank:2, source:'Feat: '+featName, id:0});
-            }
-            if (feat.length > 0 && feat[0].gainLore) {
-                this.add_LoreChoice(level, {available:1, increases:[], loreName:"", loreDesc:"", source:'Feat: '+featName, id:0});
-            }
+    take_Feat(characterService: CharacterService, featName: string, taken: boolean, level: Level, choice: FeatChoice, locked: boolean) {
+        if (taken) {
+            choice.feats.push({"name":featName, "source":choice.source, "locked":locked, "sourceId":choice.id});
+            characterService.process_Feat(featName, level, taken);
         } else {
-            let oldFeat = level.feats.filter(feat => feat.name == featName && feat.source == source)[0];
-            level.feats = level.feats.filter(feat => feat !== oldFeat);
-            let feat = characterService.get_Feats(featName);
-            if (feat.length > 0 && feat[0].increase) {
-                let oldSkillChoice = level.skillChoices.filter(choice => choice.source = 'Feat: '+featName)[0];
-                this.increase_Skill(characterService, feat[0].increase, false, oldSkillChoice, true)
-            }
-            if (source == "Level") {
-                switch (type) {
-                    case "General":
-                        level.generalFeats_applied -= 1;
-                        break;
-                    case "Class":
-                        level.classFeats_applied -= 1;
-                        break;
-                    case "Skill":
-                        level.skillFeats_applied -= 1;
-                        break;
-                    case "Ancestry":
-                        level.ancestryFeats_applied -= 1;
-                        break;
-                }
-            }
-            if (feat.length > 0 && feat[0].gainAncestryFeat) {
-                characterService.get_Character().class.levels.filter(level => level.number == 1)[0].ancestryFeats_available -= 1;
-            }
-            if (feat.length > 0 && feat[0].gainGeneralFeat) {
-                characterService.get_Character().class.levels.filter(level => level.number == 1)[0].generalFeats_available -= 1;
-            }
-            if (feat.length > 0 && feat[0].gainClassFeat) {
-                characterService.get_Character().class.levels.filter(level => level.number == 1)[0].classFeats_available -= 1;
-            }
-            if (feat.length > 0 && feat[0].gainSkillTraining) {
-                let oldChoices = level.skillChoices.filter(choice => choice.source == 'Feat: '+featName);
-                let oldChoice = oldChoices[oldChoices.length-1];
-                level.skillChoices = level.skillChoices.filter(choice => !(choice.source == oldChoice.source && choice.id == oldChoice.id))
-            }
-            if (feat.length > 0 && feat[0].gainLore) {
-                let oldChoices = level.loreChoices.filter(choice => choice.source == 'Feat: '+featName);
-                let oldChoice = oldChoices[oldChoices.length-1];
-                if (oldChoice.loreName) {
-                    this.remove_Lore(characterService, level, oldChoice);
-                }
-                level.loreChoices = level.loreChoices.filter(choice => !(choice.source == oldChoice.source && choice.id == oldChoice.id));
-            }
+            let a = choice.feats;
+            a.splice(a.indexOf(a.filter(feat => 
+                feat.name == featName &&
+                feat.locked == locked
+            )[0]), 1)
+            characterService.process_Feat(featName, level, taken);
         }
         this.set_Changed(characterService);
     }
-    remove_Lore(characterService: CharacterService, level: Level, source: LoreChoice) {
+    remove_Lore(characterService: CharacterService, source: LoreChoice) {
         let loreSkills: Skill[] = [];
         let loreFeats: Feat[] = [];
         loreSkills.push(...characterService.get_Character().customSkills.filter(skill => skill.name == 'Lore: '+source.loreName));
@@ -228,11 +166,11 @@ export class Character {
                 characterService.remove_CustomSkill(loreSkill);
             })
         }
-        /*if (loreFeats.length) {
+        if (loreFeats.length) {
             loreFeats.forEach(loreFeat => {
                 characterService.remove_CustomFeat(loreFeat);
             })
-        }*/
+        }
         //Remove the original Lore training
         characterService.get_Character().increase_Skill(characterService, 'Lore: '+source.loreName, false, source, true);
         //Go through all levels and remove skill increases for this lore from their respective sources
@@ -243,9 +181,17 @@ export class Character {
         });
         characterService.set_Changed();
     }
-    add_Lore(characterService: CharacterService, level: Level, source: LoreChoice) {
+    add_Lore(characterService: CharacterService, source: LoreChoice) {
         characterService.add_CustomSkill('Lore: '+source.loreName, "Skill", "Intelligence");
         characterService.get_Character().increase_Skill(characterService, 'Lore: '+source.loreName, true, source, true)
-        //characterService.add_LoreFeat('Lore: '+this.background.loreName, 'Background');
+        characterService.get_Feats().filter(feat => feat.lorebase).forEach(lorebase =>{
+            let newLength = characterService.add_CustomFeat(lorebase);
+            let newFeat = characterService.get_Character().customFeats[newLength -1];
+            newFeat.name = newFeat.name.replace('Lore', 'Lore: '+source.loreName);
+            newFeat.skillreq = newFeat.skillreq.replace('Lore', 'Lore: '+source.loreName);
+            newFeat.showon = newFeat.showon.replace('Lore', 'Lore: '+source.loreName);
+            newFeat.lorebase = false;
+        })
+        characterService.set_Changed();
     }
 }
